+++ adapters/Chartboost/ChartboostAdapter/GADMAdapterChartboostRewardedAd.m
@@ -14,10 +14,10 @@
 
 #import "GADMAdapterChartboostRewardedAd.h"
 
-#if __has_include(<Chartboost/Chartboost+Mediation.h>)
-#import <Chartboost/Chartboost+Mediation.h>
+#if __has_include(<ChartboostSDK/ChartboostSDK.h>)
+#import <ChartboostSDK/ChartboostSDK.h>
 #else
-#import "Chartboost+Mediation.h"
+#import "ChartboostSDK.h"
 #endif
 
 #include <stdatomic.h>
@@ -25,7 +25,6 @@
 #import "GADMAdapterChartboostConstants.h"
 #import "GADMAdapterChartboostUtils.h"
 #import "GADMChartboostError.h"
-#import "GADMChartboostExtras.h"
 
 @interface GADMAdapterChartboostRewardedAd () <CHBRewardedDelegate>
 @end
@@ -97,26 +96,20 @@
 
   NSString *adLocation = GADMAdapterChartboostLocationFromAdConfiguration(_adConfig);
   GADMAdapterChartboostRewardedAd *weakSelf = self;
-  [Chartboost startWithAppId:appID
+  [Chartboost startWithAppID:appID
                 appSignature:appSignature
-                  completion:^(BOOL success) {
+                  completion:^(CHBStartError *cbError) {
                     GADMAdapterChartboostRewardedAd *strongSelf = weakSelf;
                     if (!strongSelf) {
                       return;
                     }
 
-                    if (!success) {
+                    if (cbError) {
                       NSError *error = GADMAdapterChartboostErrorWithCodeAndDescription(
                           GADMAdapterChartboostErrorInitializationFailure,
                           @"Chartboost SDK initialization failed.");
                       strongSelf->_completionHandler(nil, error);
                       return;
-                    }
-
-                    GADMChartboostExtras *extras = strongSelf->_adConfig.extras;
-                    if (extras) {
-                      [Chartboost setFramework:extras.framework
-                                   withVersion:extras.frameworkVersion];
                     }
 
                     CHBMediation *mediation = GADMAdapterChartboostMediation();
@@ -168,11 +161,7 @@
 
 - (void)didEarnReward:(CHBRewardEvent *)event {
   [_adEventDelegate didEndVideo];
-
-  /// Chartboost doesn't provide access to the reward type.
-  NSDecimalNumber *rewardValue = [[NSDecimalNumber alloc] initWithInteger:event.reward];
-  GADAdReward *adReward = [[GADAdReward alloc] initWithRewardType:@"" rewardAmount:rewardValue];
-  [_adEventDelegate didRewardUserWithReward:adReward];
+  [_adEventDelegate didRewardUser];
 }
 
 - (void)didClickAd:(CHBClickEvent *)event error:(CHBClickError *)error {