+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiRewardedAd.m
@@ -202,12 +202,8 @@
 
 - (void)interstitial:(nonnull IMInterstitial *)interstitial
     rewardActionCompletedWithRewards:(nonnull NSDictionary *)rewards {
-  NSString *key = rewards.allKeys.firstObject;
-  if (key) {
-    GADAdReward *reward = [[GADAdReward alloc] initWithRewardType:key rewardAmount:rewards[key]];
-    [_adEventDelegate didRewardUserWithReward:reward];
-  }
   [_adEventDelegate didEndVideo];
+  [_adEventDelegate didRewardUser];
 }
 
 @end