+++ adapters/Maio/MaioAdapter/Bidding/GADRTBMaioRewardedAd.m
@@ -123,10 +123,7 @@
 }
 
 - (void)didReward:(MaioRewarded *)ad reward:(RewardData *)reward {
-  GADAdReward *gReward = [[GADAdReward alloc] initWithRewardType:reward.value
-                                                    rewardAmount:[NSDecimalNumber one]];
-
-  [_adEventDelegate didRewardUserWithReward:gReward];
+  [_adEventDelegate didRewardUser];
 }
 
 @end