+++ adapters/MoPub/MoPubAdapter/GADMAdapterMoPub.h
@@ -1,6 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <GoogleMobileAds/GoogleMobileAds.h>
-
-@interface GADMAdapterMoPub : NSObject <GADMAdNetworkAdapter>
-
-@end