+++ adapters/Vungle/VungleAdapter/GADMAdapterVungleInterstitial.m
@@ -56,6 +56,7 @@
 
 - (void)getBannerWithSize:(GADAdSize)adSize {
   id<GADMAdNetworkConnector> strongConnector = _connector;
+  [[GADMAdapterVungleRouter sharedInstance] setCOPPAStatus:strongConnector.childDirectedTreatment];
   _bannerAd = [[GADMAdapterVungleBanner alloc] initWithGADMAdNetworkConnector:strongConnector
                                                                       adapter:self];
   [_bannerAd getBannerWithSize:adSize];
@@ -65,6 +66,7 @@
 
 - (void)getInterstitial {
   id<GADMAdNetworkConnector> strongConnector = _connector;
+  [[GADMAdapterVungleRouter sharedInstance] setCOPPAStatus:strongConnector.childDirectedTreatment];
   self.desiredPlacement = [GADMAdapterVungleUtils findPlacement:[strongConnector credentials]
                                                   networkExtras:[strongConnector networkExtras]];
   if (!self.desiredPlacement.length) {
@@ -75,8 +77,8 @@
     return;
   }
 
-  VungleSDK *sdk = [VungleSDK sharedSDK];
-  if ([[GADMAdapterVungleRouter sharedInstance] hasDelegateForPlacementID:self.desiredPlacement]) {
+  VungleSDK *sdk = VungleSDK.sharedSDK;
+  if ([GADMAdapterVungleRouter.sharedInstance hasDelegateForPlacementID:self.desiredPlacement]) {
     NSError *error = GADMAdapterVungleErrorWithCodeAndDescription(
         GADMAdapterVungleErrorAdAlreadyLoaded,
         @"Only a maximum of one ad per placement can be requested from Vungle.");
@@ -96,14 +98,14 @@
     [strongConnector adapter:self didFailAd:error];
     return;
   }
-  [[GADMAdapterVungleRouter sharedInstance] initWithAppId:appID delegate:self];
+  [GADMAdapterVungleRouter.sharedInstance initWithAppId:appID delegate:self];
 }
 
 - (void)stopBeingDelegate {
   if (_bannerAd) {
     [_bannerAd cleanUp];
   } else {
-    [[GADMAdapterVungleRouter sharedInstance] removeDelegate:self];
+    [GADMAdapterVungleRouter.sharedInstance removeDelegate:self];
   }
 
   _connector = nil;
@@ -116,10 +118,10 @@
 - (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
   NSError *error = nil;
   id<GADMAdNetworkConnector> strongConnector = _connector;
-  if (![[GADMAdapterVungleRouter sharedInstance] playAd:rootViewController
-                                               delegate:self
-                                                 extras:[strongConnector networkExtras]
-                                                  error:&error]) {
+  if (![GADMAdapterVungleRouter.sharedInstance playAd:rootViewController
+                                             delegate:self
+                                               extras:[strongConnector networkExtras]
+                                                error:&error]) {
     // Ad not playable.
     if (error) {
       NSLog(@"Vungle Ad Playability returned an error: %@", error.localizedDescription);
@@ -133,8 +135,8 @@
 #pragma mark - Private methods
 
 - (void)loadAd {
-  NSError *error = [[GADMAdapterVungleRouter sharedInstance] loadAd:self.desiredPlacement
-                                                       withDelegate:self];
+  NSError *error = [GADMAdapterVungleRouter.sharedInstance loadAd:self.desiredPlacement
+                                                     withDelegate:self];
   if (error) {
     [_connector adapter:self didFailAd:error];
   }
@@ -146,8 +148,8 @@
 @synthesize isAdLoaded;
 
 - (nullable NSString *)bidResponse {
-    // This is the waterfall interstitial section. It won't have a bid response
-    return nil;
+  // This is the waterfall interstitial section. It won't have a bid response.
+  return nil;
 }
 
 - (void)initialized:(BOOL)isSuccess error:(nullable NSError *)error {
@@ -163,7 +165,7 @@
     // Already invoked an ad load callback.
     return;
   }
-    self.isAdLoaded = YES;
+  self.isAdLoaded = YES;
 
   [_connector adapterDidReceiveInterstitial:self];
 }
@@ -179,6 +181,10 @@
 
 - (void)willShowAd {
   [_connector adapterWillPresentInterstitial:self];
+}
+
+- (void)didShowAd {
+  // Do nothing.
 }
 
 - (void)didViewAd {