+++ adapters/AdColony/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## AdColony iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
+#### [Version 4.9.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/adcolony/AdColonyAdapter-4.9.0.0.zip)
+- Verified compatibility with AdColony SDK 4.9.0.
+
+Built and tested with
+- Google Mobile Ads SDK version 9.5.0.
+- AdColony SDK version 4.9.0.
+
 #### [Version 4.8.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/adcolony/AdColonyAdapter-4.8.0.0.zip)
 - Verified compatibility with AdColony SDK 4.8.0.