+++ adapters/AppLovin/CHANGELOG.md
@@ -1,6 +1,27 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
-#### Version 11.4.0.0 (In progress)
+#### [Version 11.4.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.4.3.0.zip)
+- Verified compatibility with AppLovin SDK 11.4.3.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.6.0.
+- AppLovin SDK version 11.4.3.
+
+#### [Version 11.4.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.4.2.0.zip)
+- Verified compatibility with AppLovin SDK 11.4.2.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.5.0.
+- AppLovin SDK version 11.4.2.
+
+#### [Version 11.4.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.4.1.0.zip)
+- Verified compatibility with AppLovin SDK 11.4.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.5.0.
+- AppLovin SDK version 11.4.1.
+
+#### [Version 11.4.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.4.0.0.zip)
 - Verified compatibility with AppLovin SDK 11.4.0.
 
 Built and tested with: