+++ adapters/Maio/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## maio iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
+#### [Version 1.6.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/maio/MaioAdapter-1.6.2.0.zip)
+- Verified compatibility with maio SDK version 1.6.2 and maioOB SDK version 2.0.0-alpha.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.5.0.
+- Maio SDK version 1.6.2.
+- MaioOB SDK version 2.0.0-alpha.
+
 #### [Version 1.6.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/maio/MaioAdapter-1.6.1.0.zip)
 - Verified compatibility with maio SDK version 1.6.1 and maioOB SDK version 2.0.0-alpha.