+++ adapters/Pangle/CHANGELOG.md
@@ -1,14 +1,66 @@
 ## Pangle iOS Mediation Adapter Changelog
 
+#### [Version 4.6.1.8.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.6.1.8.0.zip)
+- Verified compatibility with Pangle SDK version 4.6.1.8.
+- Added bidding support for native ad format.
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.9.0.
+- Pangle SDK version 4.6.1.8.
+
+#### [Version 4.5.2.8.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.2.8.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.2.8.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.7.0.
+- Pangle SDK version 4.5.2.8.
+
+#### [Version 4.5.2.7.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.2.7.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.2.7.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.7.0.
+- Pangle SDK version 4.5.2.7.
+
+#### [Version 4.5.2.5.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.2.5.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.2.5.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.6.0.
+- Pangle SDK version 4.5.2.5.
+
+#### [Version 4.5.2.4.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.2.4.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.2.4.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.6.0.
+- Pangle SDK version 4.5.2.4.
+
+#### [Version 4.5.2.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.2.3.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.2.3.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.6.0.
+- Pangle SDK version 4.5.2.3.
+
+#### [Version 4.5.1.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.5.1.3.0.zip)
+- Verified compatibility with Pangle SDK version 4.5.1.3.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.5.0.
+- Pangle SDK version 4.5.1.3.
+
 #### [Version 4.3.1.9.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.3.1.9.0.zip)
-- Verified compatibility with Pange SDK version 4.3.1.9.
+- Verified compatibility with Pangle SDK version 4.3.1.9.
 
 Built and tested with:
 - Google Mobile Ads SDK version 9.3.0.
 - Pangle SDK version 4.3.1.9.
 
 #### [Version 4.3.0.5.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-4.3.0.5.0.zip)
-- Verified compatibility with Pange SDK version 4.3.0.5.
+- Verified compatibility with Pangle SDK version 4.3.0.5.
 
 Built and tested with:
 - Google Mobile Ads SDK version 9.2.0.