+++ adapters/Snap/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Snap iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
 #### [Version 2.1.0.1](https://dl.google.com/googleadmobadssdk/mediation/ios/snap/SnapAdapter-2.1.0.1.zip)
 - Verified compatibility with Google Mobile Ads SDK version 9.0.0.
 - Now requires Google Mobile Ads SDK version 9.0.0 or higher.