+++ adapters/Tapjoy/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Tapjoy iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
 #### [Version 12.10.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/tapjoy/TapjoyAdapter-12.10.0.0.zip)
 - Verified compatibility with Tapjoy SDK 12.10.0.