+++ adapters/Unity/CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Unity Ads iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to use the `didRewardUser` API.
+- Now requires Google Mobile Ads SDK version 9.8.0 or higher.
+
+#### [Version 4.3.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.3.0.0.zip)
+- Verified compatibility with Unity Ads SDK 4.3.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 9.8.0.
+- Unity Ads SDK version 4.3.0.
+
 #### [Version 4.2.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.2.1.0.zip)
 - Verified compatibility with Unity Ads SDK 4.2.1.